nl2sql_next/requirements.txt
# Core pipeline
openai
pydantic>=2.0
python-dotenv

# Database
pyodbc
azure-identity

# API
fastapi
uvicorn

# Data generation (optional, for seeding)
faker
numpy
pandas
